TreeContractor

Documentation for TreeContractor.

TreeContractor.FullCompress

TreeContractor.LocalCompress

TreeContractor.canonicalize!

TreeContractor.FullCompress — Type

FullCompress <: CompressAlgorithm

FullCompress algorithm for compressing an LabeledMPS. It takes environment tensors into account and performs global SVDs, so the precision is guaranteed.

In the scenario of applying MPO to LabeledMPS, it is also called Density Matrix algorithm, which compresses the network of LabeledMPS with MPO into a single LabeledMPS.

References

https://tensornetwork.org/mps/index.html#compression

https://tensornetwork.org/mps/algorithms/denmatmpomps/

TreeContractor.LocalCompress — Type

LocalCompress <: CompressAlgorithm

LocalCompress algorithm for compressing an LabeledMPS. It performs local SVDs and updates the LabeledMPS tensors serially, so the precision is not guaranteed.

In the scenario of applying MPO to LabeledMPS, it is also called Zipup algorithm, which compresses the network of LabeledMPS with MPO into a single LabeledMPS.

References

https://tensornetwork.org/mps/algorithms/zipupmpo/

TreeContractor.canonicalize! — Method

canonicalize!(mps::LabeledMPS, i::Int; atol::Real=1e-12, maxdim::Int=typemax(Int))

Canonicalize the LabeledMPS, with the canonical center at site i. If the LabeledMPS is already canonicalized, move the center to site i.

Arguments

mps::LabeledMPS: the LabeledMPS to canonicalize

i::Int: the site index of the canonical center

atol::Real=1e-12: the truncation error

maxdim::Int=typemax(Int): the maximum bond dimension for truncation